requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
Pillow
psycopg2-binary
gunicorn
whitenoise
python-decouple
cloudinary  # optional for video hosting
boto3       # if using AWS S3